fix: fix relayoutboundary and position box.
