+++ webf/lib/rendering.dart
@@ -13,6 +13,7 @@ export 'src/rendering/box_model.dart';
 export 'src/rendering/replaced.dart';
 export 'src/rendering/sliver_list.dart';
 export 'src/rendering/preferred_size.dart';
+export 'src/rendering/render_position_holder.dart';
 export 'src/rendering/text.dart';
 export 'src/rendering/text_control.dart';
 export 'src/rendering/content_visibility.dart';
