+++ webf/lib/src/rendering/preferred_size.dart
@@ -42,44 +42,3 @@ class RenderPreferredSize extends RenderProxyBox {
     properties.add(DiagnosticsProperty<Size>('preferredSize', _preferredSize, missingIfNull: true));
   }
 }
-
-/// A placeholder for positioned RenderBox
-class RenderPositionPlaceholder extends RenderPreferredSize {
-  RenderPositionPlaceholder({
-    required Size preferredSize,
-    RenderBox? child,
-  }) : super(preferredSize: preferredSize, child: child);
-
-  // Real position of this renderBox.
-  RenderBoxModel? positioned;
-
-  // Box size equals to RenderBox.size to avoid flutter complain when read size property.
-  Size? _boxSize;
-
-  Size? get boxSize {
-    assert(_boxSize != null, 'box does not have laid out.');
-    return _boxSize;
-  }
-
-  @override
-  set size(Size value) {
-    _boxSize = value;
-    super.size = value;
-  }
-
-  @override
-  bool hitTest(BoxHitTestResult result, {Offset? position}) {
-    return false;
-  }
-
-  // Get the layout offset of renderObject to its ancestor which does not include the paint offset
-  // such as scroll or transform.
-  Offset getOffsetToAncestor(Offset point, RenderObject ancestor, {bool excludeScrollOffset = false}) {
-    return MatrixUtils.transformPoint(
-        getLayoutTransformTo(this, ancestor, excludeScrollOffset: excludeScrollOffset), point);
-  }
-}
-
-bool isPositionPlaceholder(RenderBox box) {
-  return box is RenderPositionPlaceholder;
-}
